Refactor payment and wallet system to use double amounts
- Updated Wallet entity to use Double for balance
- Refactored WalletService to handle double amounts for credit, debit, lock, and unlock
- Updated PaymentService to work with double amounts and credit owner wallet correctly
- Added PaymentController and WebhookController for creating invoices and handling payment callbacks
- Updated UserRepository and WalletRepository with proper query methods
- Modified User entity and Notification entity as needed
- Adjusted application properties and pom.xml as required
- Updated UserController and NotificationUtil to support new wallet/payment changes

+++ src/main/java/bullbear/app/entity/user/Notification.java
@@ -12,9 +12,9 @@ public class Notification {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer notificationId;
+    private Long notificationId;
 
-    private Integer userId;
+    private Long userId;
 
     private String type; // e.g., TRANSACTION, SYSTEM, COMMISSION
